Dementia Prediction Using Machine Learning
Dementia is a dangerous condition that affects the brain's ability to think, communicate, and remember, according to Statista. It affects more than 50 million people worldwide. By 2050, it is projected to reach more than 152 million. The World Health Organization (WHO) projects that there are 10 million new cases each year, 60 to 70 percent of which proceed to Alzheimer's disease.